# swagger.yaml
basePath: /
consumes:
- text/plain
definitions:
...
  addReplayDataReq:
    properties:
      AddMsgCateID:
        format: uint32
        type: integer
      AngX:
        format: float
        type: number
      AngY:
        format: float
        type: number
      AngZ:
        format: float
        type: number
      BlockID:
        format: uint32
        type: integer
      CharacterID:
        type: string
      Data:
        type: string
      MainMsgID:
        format: uint32
        type: integer
      MsgID:
        format: uint32
        type: integer
      PosX:
        format: float
        type: number
      PosY:
        format: float
        type: number
      PosZ:
        format: float
        type: number
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  addSosDataReq:
    properties:
      AddMsgCateID:
        format: uint32
        type: integer
      AngX:
        format: float
        type: number
      AngY:
        format: float
        type: number
      AngZ:
        format: float
        type: number
      Black:
        format: uint8
        type: integer
      BlockID:
        format: uint32
        type: integer
      CharacterID:
        type: string
      MainMsgID:
        format: uint32
        type: integer
      MsgID:
        format: uint32
        type: integer
      PlayerInfo:
        type: string
      PlayerLevel:
        format: uint32
        type: integer
      PosX:
        format: float
        type: number
      PosY:
        format: float
        type: number
      PosZ:
        format: float
        type: number
      QWCLR:
        format: uint32
        type: integer
      QWCWB:
        format: uint32
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  checkSosDataReq:
    properties:
      CharacterID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  getSosDataReq:
    properties:
      Black:
        format: int64
        type: integer
      BlackMax:
        format: int64
        type: integer
      BlackMin:
        format: int64
        type: integer
      BlockID:
        format: uint32
        type: integer
      Invate:
        format: int64
        type: integer
      InvateMax:
        format: int64
        type: integer
      InvateMin:
        format: int64
        type: integer
      MaxSOSNum:
        format: int64
        type: integer
      PlayerLevelMax:
        format: int64
        type: integer
      PlayerLevelMin:
        format: int64
        type: integer
      SOSList:
        type: string
      SOSNum:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  replayDataReq:
    properties:
      GhostID:
        format: uint32
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  replayListReq:
    properties:
      BlockID:
        format: uint32
        type: integer
      ReplayNum:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  summonBlackGhostReq:
    properties:
      NPRoomID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  summonOtherCharacterReq:
    properties:
      GhostID:
        format: int32
        type: integer
      NPRoomID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
info:
  description: The boostrap and game API endpoints for Demon's Souls
  title: Demon's Souls API
  version: 1.0.0
paths:
...
  /cgi-bin/addReplayData.spd:
    post:
      consumes:
      - text/plain
      description: Adds replay data for the given character
      operationId: addReplayDataHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/addReplayDataReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Add replay data
      tags:
      - replays
  /cgi-bin/addSosData.spd:
    post:
      consumes:
      - text/plain
      description: Adds an SOS message for a given character
      operationId: addSosDataHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/addSosDataReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Add SOS message
      tags:
      - sos
...
  /cgi-bin/checkSosData.spd:
    post:
      consumes:
      - text/plain
      description: |-
        Checks for a matching SOS for the given character, part of the matchmaking
        process.
      operationId: checkSosDataHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/checkSosDataReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Check SOS message
      tags:
      - sos
...
  /cgi-bin/getReplayData.spd:
    post:
      consumes:
      - text/plain
      description: Returns a single replay's data
      operationId: getReplayDataHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/replayDataReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get replay data
      tags:
      - replays
  /cgi-bin/getReplayList.spd:
    post:
      consumes:
      - text/plain
      description: Returns a list of available replays for an area of the game
      operationId: replayListHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/replayListReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: List replays
      tags:
      - replays
  /cgi-bin/getSosData.spd:
    post:
      consumes:
      - text/plain
      description: Returns a list of SOS messages for a given area of the game
      operationId: getSosDataHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/getSosDataReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get SOS message
      tags:
      - sos
...
  /cgi-bin/summonBlackGhost.spd:
    post:
      consumes:
      - text/plain
      operationId: summonBlackGhostHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/summonBlackGhostReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Summon black ghost
      tags:
      - sos
  /cgi-bin/summonOtherCharacter.spd:
    post:
      consumes:
      - text/plain
      description: |-
        Summon another character to assist via a multiplayer session, part of the
        matchmaking process.
      operationId: summonCharacterHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/summonOtherCharacterReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Summon other character
      tags:
      - sos
...
produces:
- text/plain
schemes:
- http
swagger: "2.0"